requirements.txt
requests
# beautifulsoup4
pandas
numpy
tqdm
scikit-learn
faiss-cpu
sentence-transformers
python-dotenv
kaggle
huggingface_hub
streamlit
